feat: configure Jest and RTL
