+++ package.json
@@ -10,7 +10,9 @@
     "deploying": "gh-pages -d dist",
     "predeploy": "npm run build",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "jest",
+    "test:coverage": "jest --coverage --collectCoverageFrom='src/**/*.{js,jsx}'"
   },
   "dependencies": {
     "@emotion/react": "^11.11.1",
@@ -22,6 +24,8 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.1.6",
+    "@testing-library/react": "^14.1.2",
     "@types/node": "20.10.5",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
@@ -30,6 +34,8 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
     "gh-pages": "^6.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "typescript": "5.3.3"
   }
 }
